backend/api/requirements.txt
Flask
Flask-JWT-Extended==4.5.3
Flask-Cors
Flask-SocketIO
python-dotenv
Werkzeug
pymongo
opencv-python
dlib
scipy
imutils
